update_4.1

+++ app/src/main/java/com/sbdevs/booksonlineseller/otherclass/FixedPriceClass.kt
@@ -1,7 +1,9 @@
 package com.sbdevs.booksonlineseller.otherclass
 
+import java.math.BigDecimal
+
 class FixedPriceClass {
     companion object{
-        val pickupCharge:Long = 30L
+        val pickupCharge:BigDecimal = BigDecimal(30)
     }
 }
